requirements.txt
anytree==2.8.0
matplotlib==3.5.1
networkx==2.6.3
numpy==1.22.2
opencv-python==4.5.5.62
Pillow==9.0.1
pygame==2.1.2
scikit-image==0.19.2
scipy==1.8.0
Kivy~=2.1.0
pyFluidSynth~=1.3.0
PyAutoGUI~=0.9.53
PyGetWindow~=0.0.9
mss~=6.1.0
pyttsx3~=2.90
tzlocal~=4.2
shapely~=2.0.1
openvr~=1.23.701